disconnect sessions after drop tables for mysql/postgresql

+++ hugegraph-mysql/src/main/java/com/baidu/hugegraph/backend/store/mysql/MysqlStore.java
@@ -168,14 +168,24 @@ public abstract class MysqlStore extends AbstractBackendStore<Session> {
     }
 
     @Override
-    public void clear() {
+    public void clear(boolean clearSpace) {
         // Check connected
         this.checkClusterConnected();
 
         if (this.sessions.existsDatabase()) {
-            this.checkOpened();
-            this.clearTables();
-            this.sessions.dropDatabase();
+            if (!clearSpace) {
+                this.checkOpened();
+                this.clearTables();
+                /*
+                 * Disconnect connections for following database drop.
+                 * Connections will be auto reconnected if not drop database
+                 * in next step, but never do this operation because database
+                 * might be blocked in mysql or throw 'terminating' exception.
+                 */
+                this.sessions.resetConnections();
+            } else {
+                this.sessions.dropDatabase();
+            }
         }
 
         LOG.debug("Store cleared: {}", this.store);
